+++ package.json
@@ -4,16 +4,23 @@
   "description": "",
   "main": "./src/main/main.js",
   "scripts": {
-    "start": "electron .",
+    "start": "npm run build:server && electron .",
     "build": "webpack --config webpack.config.js",
-    "dev": "concurrently \"npx cross-env NODE_ENV=development npx webpack serve --config webpack.config.js\" \"npx cross-env NODE_ENV=development npx nodemon --watch src/main/main.js --exec electron .\""
+    "dev": "concurrently \"npx cross-env NODE_ENV=development npx webpack serve --config webpack.config.js\" \"npx cross-env NODE_ENV=development npx nodemon --watch src/main/main.js --exec electron .\"",
+    "build:server": "npx tsc",
+    "dev:server": "npx tsc --watch"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@babel/preset-env": "^7.24.7",
-    "@babel/preset-react": "^7.24.7",
+    "@babel/plugin-transform-typescript": "^7.25.7",
+    "@babel/preset-env": "^7.25.7",
+    "@babel/preset-react": "^7.25.7",
+    "@babel/preset-typescript": "^7.25.7",
+    "@types/express": "^4.17.21",
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.0",
     "babel-loader": "^9.1.3",
     "concurrently": "^8.2.2",
     "cross-env": "^7.0.3",
@@ -24,6 +31,8 @@
     "jest": "^29.7.0",
     "nodemon": "^3.1.4",
     "tailwindcss": "^3.4.10",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.6.2",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.4"
   },
@@ -60,6 +69,7 @@
     "redux": "^5.0.1",
     "simple-oauth2": "^5.1.0",
     "style-loader": "^4.0.0",
+    "ts-node": "^10.9.2",
     "webpack": "^5.92.1"
   }
 }
