+++ src/renderer/index.js
@@ -9,4 +9,3 @@ const root = createRoot(document.getElementById('root'));
 root.render(
   <App />
 );
-
