--- src/server/controller/user/userController.js
@@ -1,93 +1,0 @@
-const jwt = require('jsonwebtoken');
-const secretKey = 'cats';
-const User = require('../../models/user');
-// const { Next } = require('react-bootstrap/esm/PageItem');
-
-const userController = {
-  // Middleware to create a new user 
-  async createUser(req, res, next) {
-
-    const { username, email, password } = req.body;
-
-    if (!username || !email || !password) {
-      return res.status(500).send('Error: missing information');
-    };
-
-    try {
-      // Save user to database 
-      const newUser = new User({ username, email, password });
-      await newUser.save();
-
-      // Store user's id in res.locals
-      // console.log('newUser id is: ', newUser._id.toString());
-      res.locals.userId = newUser._id.toString();
-     
-      return next();
-
-    } catch (err) {
-      return res.status(500).send(`Error in create user controller: ${err}`);
-    }
-  },
-
-  // Middleware to login 
-  async login(req, res, next) {
-
-    console.log('login middleware hit')
-
-    const { username, password } = req.body;
-
-    if (!username || !password) {
-      return res.status(400).send('Error: missing username or password');
-    }
-
-    try {
-      console.log('in login try block')
-      const user = await User.findOne({ username, password });
-      console.log('user found? ', user)
-
-      const userId = user._id;
-      console.log('userId is ', userId.toString());
-      res.locals.userId = userId.toString();
-
-      if (!user) {
-        return res.status(404).send('User not found');
-      }
-
-      // DELETE THESE 2 LINES? 
-      // const token = jwt.sign({ id: user._id, username: user.username }, secretKey, { expiresIn: '1h' });
-      // res.cookie('token', username, { httpOnly: true, secure: true });
-
-      // return res.status(200).send(token);
-      return next();
-
-    } catch (err) {
-      return res.status(500).send(`Error in login controller: ${err}`);
-    }
-  },
-
-  async saveToken(req, res, next) {
-    // console.log("req.body", req.body)
-    const { access_token } = req.body;
-
-    if (!access_token) {
-      return res.status(500).send('Error: missing information in save toke');
-    }
-
-    try {
-      const newUser = new User({ access_token });
-      await newUser.save();
-
-      // const token = jwt.sign({ id: newUser._id, username: newUser.username }, secretKey, { expiresIn: '1h' });
-      // res.cookie('token', username, { httpOnly: true, secure: true });
-
-      // return res.status(200).send(newUser);
-
-      return next();
-
-    } catch (err) {
-      return res.status(500).send(`Error in create user controller: ${err}`);
-    }
-  }
-};
-
-module.exports = userController;
