+++ src/server/routes/awsRouter.ts
@@ -2,7 +2,7 @@
 // Imports and Configuration
 // -----------------------------------------
 
-const express = require('express');
+import express, { Request, Response, NextFunction } from 'express';
 const path = require('path');
 const router = express.Router();
 
@@ -15,26 +15,26 @@ const getLambdaMetrics = require('../controller/aws/metricsController.js');
 // -----------------------------------------
 
 // Route to add AWS credential information
-router.post('/credential/add',awsCredential.addAWSCredential, (req, res) => {
+router.post('/credential/add',awsCredential.addAWSCredential, (req: Request, res: Response) => {
   return res.status(200).send('message');
 });
 
 // Route to get AWS lambda's invocation count 
-router.get('/metric/invocation', getLambdaMetrics.getInvocationCount, (req, res) => {
+router.get('/metric/invocation', getLambdaMetrics.getInvocationCount, (req: Request, res: Response) => {
   return res.status(200).json(res.locals.invocationData);
 });
 
 // Route to get AWS lambda's error count 
-router.get('/metric/error', getLambdaMetrics.getErrorCount, (req, res) => {
+router.get('/metric/error', getLambdaMetrics.getErrorCount, (req: Request, res: Response) => {
   return res.status(200).json(res.locals.errorData);
 });
 
 // Route to get AWS lambda's throttle count 
-router.get('/metric/throttle', getLambdaMetrics.getThrottleCount, (req, res) => {
+router.get('/metric/throttle', getLambdaMetrics.getThrottleCount, (req: Request, res: Response) => {
   return res.status(200).json(res.locals.throttleData);
 });
 
-router.get('/metric/duration', getLambdaMetrics.getDuration, (req, res) => {
+router.get('/metric/duration', getLambdaMetrics.getDuration, (req: Request, res: Response) => {
   return res.status(200).json(res.locals.durationData);
 });
