+++ src/server/types.ts
@@ -1,0 +1,5 @@
+export type CustomError = {
+  log: string,
+  status: number,
+  message: string | {err: string}
+};
